moved isad file import into new FileImporter class
